requirements.txt
streamlit
stream-chat
openai
python-dotenv
